requirements.txt
advent-of-code-data==2.0.1
advent-of-code-ocr==1.0.0
aocd-example-parser==2023.2
beautifulsoup4==4.12.2
black==23.11.0
click==8.1.7
colorama==0.4.6
iniconfig==2.0.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parse==1.20.0
pathspec==0.11.2
Pebble==5.0.4
platformdirs==4.0.0
pluggy==1.3.0
Pygments==2.17.2
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
rich==13.7.0
scipi==3.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
tzdata==2023.3
urllib3==2.1.0
wheel==0.41.3
